requirements.txt
git+https://github.com/erroneousboat/warc3
requests
dnspython
